- Move saving emails with attachments in the database to the sending email job for preventing saving emails into the database when sending emails fail.
